introduce decode mod
