+++ ezlog-core/src/config.rs
@@ -564,7 +564,7 @@ mod tests {
             .name(String::from("all_feature"))
             .file_suffix(String::from("mmap"))
             .compress(CompressKind::ZLIB)
-            .cipher(CipherKind::AES128GCM)
+            .cipher(CipherKind::AES128GCMSIV)
             .cipher_key(key.to_vec())
             .cipher_nonce(nonce.to_vec())
             .max_size(1024)
